requirements.txt
flask
pandas
google-cloud-vision
pillow